+++ include/services.h
@@ -60,7 +60,6 @@ struct nicksvs
 	unsigned int    enforce_expiry; // expiry time
 	unsigned int    enforce_delay;  // delay for nickname enforce
 	char *          enforce_prefix; // prefix for enforcement
-	char *          cracklib_dict;  // cracklib dictionary path
 	mowgli_list_t   emailexempts;   // emails exempt from maxusers checks
 };